defs: Update method callback signature
This commit changes the callback signature used for methods. The new signature
has the same first three parameters, but now included a fourth that will
provide the callback with the request object that initiated the callback call.
In addition, the callback now behaves similarly to Java overloaded functions;
the callback can be specified with 0 to 4 parameters (positional, not keyword)
and the API will call it accordingly.

Signed-off-by: Ben Wiebe <ben.wiebe@windriver.com>