handler: Fix disconnect failing when called within a method callback
This commit fixes a bug with the handler disconnect method raising an
exception when called from a method callback. Previously, calling disconnect
would cause the callback thread to attempt to join itself. This is fixed by
checking if the thread that is to be joined is the same as the current
(caller) thread, and if so do not attempt to join.

Signed-off-by: Ben Wiebe <ben.wiebe@windriver.com>