handler: No longer timing out sent messages
Since we are using QOS 1 for MQTT, the Cloud is guaranteed to receive at least
one copy of each message as long as the device is on-line. MQTT will already
attempt to resend any messages that have not been confirmed received. This means
we don't need to worry about message timeouts explicitly.

Signed-off-by: Ben Astell <Ben.Astell@windriver.com>