relay: Add a websocket relay to the device manager
Adds a websocket relay which can be used for remote login. Thing definition
must have a remote-access method with url, host, and protocol notification
parameters. The url is for the websocket connection to a Cloud service, host and
protocol are for the socket connection to a local service, such as Telnet.

Signed-off-by: Ben Astell <Ben.Astell@windriver.com>