file io: allow files to be uploaded/downloaded to/from the global files
This commit adds a global parameter to the TR50 requests for file upload and
download. If true, the file will be uploaded to, or downloaded from, the
global file store on the cloud. The Device Manager was also updated to allow
the user to select the desired location when running file IO methods.

Signed-off-by: Ben Wiebe <ben.wiebe@windriver.com>