device manager: publish more system attributes
This commit adds additional system attributes to be published by the device
manager app. All attributes listed in US100702 are published to the cloud when
the app first starts. It should be noted that the API version is now set in
hdcpython/constants.py to allow for it to be accessed through the library as
well as by setup.py for installation.

Signed-off-by: Ben Wiebe <ben.wiebe@windriver.com>