download: return STATUS_NOT_FOUND if file isn't on the cloud
This commit modifies the file download flow so that if the file does not
exist on the cloud, STATUS_NOT_FOUND is returned instead of STATUS_FAILURE.
This aligns more closely with the file upload returns and allows for better
error handling by client apps. The OTA handler was also updated to use this
new status when checking file download status.

Signed-off-by: Ben Wiebe <ben.wiebe@windriver.com>