install: fix dependency issue V2
This commit addresses an oversight in the previous dependency fix which
resulted in the issue not actually being fixed. In this patch, the version is
only in setup.py to avoid dependencies. The device manager now reads the HDC
version from the installed package information. Unfortunately, this means that
the version can only be published if hdcpython is installed via pip or
via setup.py.

Signed-off-by: Ben Wiebe <ben.wiebe@windriver.com>