install: add Python version checking
This commit adds a check to setup.py to verify that the correct version of
Python is installed on the system. Currently, the minimum versions are 2.7.9
and 3.4, however this may change as the module evolves. If the system is not
using a version greater than or equal to these, then the installation will
fail with a notice informing the user to upgrade their Python.

Signed-off-by: Ben Wiebe <ben.wiebe@windriver.com>