device manager & osal: fix device manager-related issues
This commit fixes two issues found in the device manager.  The first was the
acknowledgement of saved messages (eg. reset agent) was not using the new
client API to do so, which resulted in calling a module that had been moved
in the "helix" restructure. In addition, when running with Python 3, the OS
name would show a byte literal indicator in the platform part.

Signed-off-by: Ben Wiebe <ben.wiebe@windriver.com>