test & lib: add more unit tests and fix minor issues in library
This commit adds more unit testing to the helix module along with a coverage
configuration file to omit tests from coverage. In addition, it fixes an issue
with the implementation of execl in the OSAL which prevented commands from
being run if they did not need any parameters. The relay loop method was also
prefixed with an underscore to follow Python convention indicating that the
method should not be called externally.

Signed-off-by: Ben Wiebe <ben.wiebe@windriver.com>