share: add basic OTA package creation tool
This commit adds a new Python script to the share directory that can be used
to create OTA packages. The tool can be run either in a headless mode (with
parameters specified) or normally which will prompt the user for values. The
OTA manifest will be generated and any specified files will be added to a zip
or tar archive.

Signed-off-by: Ben Wiebe <ben.wiebe@windriver.com>