validate_script: delete existing thing key in cloud
Instead of creating a new thing key for every new run of the test app,
i.e. in docker the device_id is generated every time, use a default
device_id if a device_id file does not exist.  Then, delete the
thing_key in the cloud so that there are not stale results.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>