handler.py: add fix for OTA status update.
During OTA, the work flow is acknowledge right away with "invoked",
then update the status later.  This way we can handle async status
updates.  In this case, the handler was sending a mailbox.ack which
was not the correct behaviour.  The cloud sees that as a failure.
Eventhough it is updated later, the status was failure.  Add logic to
the handler to check the retcode for "invoked".  If so, send an
mailbox.update instead of an ack.

Signed-off-by: Paul Barrette <paulbarrette@gmail.com>